requirements.txt
python-docx>=1.1.0
Pillow>=10.4.0
customtkinter>=5.2.2
lxml>=5.3.0
